% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testmegawrapper.R
\name{runMega}
\alias{runMega}
\title{Test of MEGA in R}
\usage{
runMega(analysis_file, data_file, out_parse = "txt",
  delete_file = FALSE)
}
\arguments{
\item{analysis_file}{Filename of MAO file generated by prototyper}

\item{data_file}{Filename of data file}

\item{out_parse}{Datatype of MEGA output}

\item{out_file}{Name of temporary output file}
}
\value{
Data from MEGA
}
\description{
Test of MEGA in R
}
\author{
Jared Knoblauch
}
